test: split up vitest setup scripts
# Conflicts:
#	mocks/msw/setup.ts
