+++ mocks/logger/setup.ts
@@ -1,0 +1,10 @@
+import { afterAll, beforeAll } from 'vitest';
+import { logger } from '../../src/utils/logger';
+
+beforeAll(() => {
+  logger.silent = true;
+});
+
+afterAll(() => {
+  logger.silent = false;
+});
