+++ mocks/msw/setup.ts
@@ -1,5 +1,5 @@
 import { afterAll, afterEach, beforeAll } from 'vitest';
-import { server } from './mocks/server';
+import { server } from './server';
 
 beforeAll(() => {
   server.listen();
